Issue-56 - committed latest changes: todo - add the title toolbar
