+++ app/build.gradle
@@ -98,7 +98,7 @@ dependencyUpdates {
 
 dependencies {
     def CONSTRAINT_LAYOUT_VERSION = '2.0.0-alpha2'
-    def MATERIAL_VERSION = '1.1.0-alpha05'
+    def MATERIAL_VERSION = '1.1.0-alpha08'
     def GRID_LAYOUT_VERSION = '1.0.0'
     def CORE_VERSION = '1.0.1'
     def MEDIA_VERSION = '1.1.0-alpha02'
@@ -118,6 +118,7 @@ dependencies {
     def TEST_RULES_VERSION = '1.1.1'
     def ESPRESSO_CORE_VERSION = '3.1.0'
     def GOOGLE_GUAVA_VERSION = '27.1-android'
+    def COORDINATOR_LAYOUT_VERSION = '1.0.0'
     def TEST_UI_AUTOMATOR_VERSION= '2.2.0'
     def FRAGMENT_TESTING_VERSION = '1.1.0-alpha07'
 
@@ -129,6 +130,7 @@ dependencies {
     implementation group: 'androidx.annotation', name: 'annotation', version: ANNOTATION_VERSION
     debugImplementation group: 'androidx.appcompat', name: 'appcompat', version: APP_COMPAT_VERSION
     implementation group: 'androidx.drawerlayout', name: 'drawerlayout', version: DRAWER_LAYOUT_VERSION
+    implementation group: 'androidx.coordinatorlayout', name: 'coordinatorlayout', version: COORDINATOR_LAYOUT_VERSION
     implementation group: 'androidx.core', name: 'core', version: CORE_VERSION
     implementation group: 'org.apache.commons', name: 'commons-io', version: COMMONS_IO_VERSION
     implementation group: 'org.apache.commons', name: 'commons-lang3', version: COMMONS_LANG_VERSION
