+++ app/src/main/res/layout/fragment_main_frame.xml
@@ -1,35 +1,66 @@
-<androidx.drawerlayout.widget.DrawerLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/drawer_layout"
-    android:fitsSystemWindows="true"
-    >
+    android:fitsSystemWindows="true">
 
     <RelativeLayout
-        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        tools:context="com.example.mike.mp3player.client.activities.MainActivity">
+        android:layout_height="match_parent">
 
-        <fragment
-            android:id="@+id/titleBarFragment"
+        <androidx.coordinatorlayout.widget.CoordinatorLayout
+            android:id="@+id/overview_coordinator_layout"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
-            class="com.example.mike.mp3player.client.views.fragments.TitleBarFragment"
-            tools:layout="@layout/fragment_title_bar"/>
-
-        <fragment
-            android:id="@+id/viewPagerFragment"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/titleBarFragment"
+            android:layout_height="match_parent"
             android:layout_above="@+id/playToolbarFragment"
-            class="com.example.mike.mp3player.client.views.fragments.viewpager.ViewPagerFragment"
-            tools:layout="@layout/fragment_view_pager"/>
+            tools:context="com.example.mike.mp3player.client.activities.MainActivity">
+
+            <com.google.android.material.appbar.AppBarLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:fitsSystemWindows="true"
+                >
+
+                <com.google.android.material.appbar.CollapsingToolbarLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="?attr/actionBarSize"
+                    app:titleTextColor="?attr/textColorPrimary"
+                    app:expandedTitleGravity="top"
+                    app:contentScrim="?attr/colorPrimary"
+                    app:layout_scrollFlags="scroll|exitUntilCollapsed|enterAlways"
+                >
+
+                        <ImageButton
+                            android:id="@+id/searchFilter"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:theme="@style/AppTheme"
+                            android:layout_gravity="end"
+                            android:layout_marginEnd="@dimen/default_gap"
+                            android:background="@android:color/transparent"
+                            android:contentDescription="@string/SEARCH_BUTTON_DESCRIPT"
+                            app:srcCompat="@drawable/ic_search_black_24dp" />
+
+                </com.google.android.material.appbar.CollapsingToolbarLayout>
+
+                <com.google.android.material.tabs.TabLayout
+                    android:id="@+id/tabs"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    app:layout_collapseMode="pin" />
+
+            </com.google.android.material.appbar.AppBarLayout>
+
+            <androidx.viewpager.widget.ViewPager
+                android:id="@+id/rootItemsPager"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_above="@+id/playToolbarFragment"
+                app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+
+        </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
         <fragment
             android:id="@+id/playToolbarFragment"
@@ -38,7 +69,6 @@
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             tools:layout="@layout/fragment_playback_toolbar" />
-
 
     </RelativeLayout>
