+++ app/src/main/res/layout/fragment_title_bar.xml
@@ -1,16 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:layout_width="match_parent"
+<com.google.android.material.appbar.AppBarLayout android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:gravity="top"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="top"
-       >
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/titleToolbar"
@@ -34,4 +27,3 @@
         </androidx.appcompat.widget.Toolbar>
 
     </com.google.android.material.appbar.AppBarLayout>
-</LinearLayout>
