--- app/src/main/res/layout/fragment_view_pager.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:background="@color/white"
-    android:layout_height="match_parent">
-
-    <androidx.viewpager.widget.ViewPager
-        android:id="@+id/rootItemsPager"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent">
-
-        <androidx.viewpager.widget.PagerTabStrip
-            android:id="@+id/pagerTabStrip"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="top"
-            android:paddingBottom="4dp"
-            android:paddingTop="4dp" />
-
-    </androidx.viewpager.widget.ViewPager>
-
-</LinearLayout>
